Reorganize components for easy styling and rename for orgainzation
