+++ client/components/NavAuth.js
@@ -6,15 +6,11 @@ export const AuthNav = React.createClass({
 
   render: function() {
     return (
-      <div className="auth-nav">
-        <img className="faunadex-logo" src="http://vignette2.wikia.nocookie.net/farmville/images/c/c2/Deep_Sea_Manatee-icon.png" />
-        <h1 className="faunadex-title">Faunadex</h1>
-        <ul className="nav-links">
-          <li className="share-new-encounter-link"><a href="/#/newencounter">Share New Encounter</a></li>
-          <li className="user-profile-link"><a href="/#/userprofile">{this.props.username}</a></li>
-          <li className="logout-link"><a onClick={this.props.signOut} href="/#/">Signout</a></li>
-        </ul>
-      </div>
+      <ul>
+        <li className="share-new-encounter-link"><a href="/#/newencounter">Share New Encounter</a></li>
+        <li className="user-profile-link"><a href="/#/userprofile">{this.props.username}</a></li>
+        <li className="logout-link"><a onClick={this.props.signOut} href="/#/">Signout</a></li>
+      </ul>
     );
   }
 });
